+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Fix layer `AnchorSize::Window` not filling the window space by default.
 * Fix `ContextCapture::NoCapture` excluding popup config.
 * Add `ResponseVar::map_response`.
 * Add `Dialog!` widget, `DIALOG` service and related types.